Fix "bug" in the data_generator
Before it was working only for batch_size < file_size, plus it was truncating the last portion of each file. Now it successfully fill in the batches with all of the data, for all batch sizes. The only portion discarded is the remaining part of the last file it loads.